requirements.txt
requests
beautifulsoup4
python-pptx
lxml
selenium
pytest